fix(plan): return long for number provider instead of int

+++ src/main/java/de/raidcraft/referrals/plan/ReferralDataExtension.java
@@ -75,7 +75,7 @@ public class ReferralDataExtension implements DataExtension {
             iconColor = Color.NONE,
             showInPlayerTable = true
     )
-    public int referralCount(UUID playerUUID) {
+    public long referralCount(UUID playerUUID) {
 
         Optional<ReferralPlayer> player = ReferralPlayer.byId(playerUUID);
         if (player.isEmpty()) return 0;
